stable_ver/v9/imgw_proxy2/imgwproj/requirements.txt
Django>=5.2,<6
djangorestframework>=3.16,<4
requests>=2.32,<3